+++ src/constants/appVersion.ts
@@ -1,1 +1,1 @@
-export const APP_VERSION = "13.297.3";
+export const APP_VERSION = "13.297.4";